Moving to monorepo, documenting using gitbook
