+++ .gitbook.yaml
@@ -1,0 +1,3 @@
+structure:
+  readme: README.md
+  summary: docs/README.md
