+++ README.md
@@ -40,72 +40,3 @@
 > will change by the final framework, so use at your own discretion.
 >
 > \- Liam
-
-## Initial plans (now severely outdated, only for reference)
-
-```js
-const myPaper = new Paper({ width, height, plugins, ...etc });
-myPaper.getPaperElement(); // Returns SVGElement
-
-// TODO:  ...
-
-INodeComponent
-INodeComponent.getNodeElement()
-INodeComponent.getParameters(): {width height}
-INodeComponent.updateProps(some prop change)
-
-myPaper.addNode({
-    id: string (unique for paper instance),
-    coords: ICoords{ // Top-left corner of the node
-        x,
-        y,
-        [z]
-    },
-    component: INodeComponent,
-    props: IComponentProps{
-        [
-          content: HTMLElement,
-            (or),
-          title: string,
-          icon: string
-      ]
-    }
-}); // Returns Promise: returns the node object if success, otherwise return error
-
-myPaper.removeNode(id);
-
-myPaper.getNode(id).updateProps({
-    ...patchedProps
-});
-
-myPaper.addEdge({
-    id: string (unique),
-    points: [ICoords],
-    component: IEdgeComponent,
-    props: IEdgeComponentProps{
-        [content, (or), title, icon]
-    },
-    source: { (either an object with id, or ICoords) },
-    target: { (either an object with id, or ICoords) }
-});
-
-// Collision detection (inside Paper) - will not trigger 'add-edge'
-// Workflow Data validation
-const myListener = myPaper.addListener('add-edge', async (event, edge) => {
-    if (await validate(edge)) {
-        try {
-            edge.updateProps(new edge or something)
-        }
-        catch(e) {
-            probably remove edge
-            // const myEdge = await myPaper.addEdge(edge);
-            // this.addEdgeToState(myEdge);
-        }
-    }
-    else {
-        myPaper.removeEdge(edge.id)
-    }
-});
-
-myListener.remove( );
-```
