+++ docs/README.md
@@ -1,0 +1,5 @@
+# Table of Contents
+
+- [Read Me](../README.md)
+- [Introduction](introduction/README.md)
+  - [Motivation](introduction/Motivation.md)
