+++ docs/introduction/Motivation.md
@@ -1,0 +1,3 @@
+# Motivation
+
+To build a robust graphing framework that doesn't use outdated dependencies.
