+++ package.json
@@ -1,10 +1,7 @@
 {
   "name": "psiagram",
-  "version": "0.1.0",
-  "author": "Liam Ross",
-  "main": "lib/index.js",
-  "typings": "lib/index.d.ts",
   "private": true,
+  "license": "MIT",
   "scripts": {
     "start": "parcel demo/index.html",
     "test": "jest --config test/jest.config.js",
