+++ packages/psiagram/components/Paper/Paper.types.ts
