+++ packages/psiagram/components/Paper/definitions/svgDefinitions.ts
