+++ test/transformers/typescript.js
